pyproject.toml
[tool.poetry]
name = "foragerpy"
version = "0.1.0"
description = "The official Python client for Forager."
authors = ["Fait Poms <faitpoms@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
tqdm = "^4.61.2"
aiohttp = "^3.7.4"
Pillow = "^8.3.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
